fix(): Удалил GeopositionRepository
